app/services/llava_service.py
"""
LLaVA模型服务
调用自研多模态模型生成医学报告
"""
from loguru import logger
from typing import Tuple
import time
import os


class LLaVAService:
    """
    LLaVA模型调用服务

    【TODO - 后端团队成员需要实现】:
    1. 实现真实的LLaVA模型推理逻辑
    2. 优化提示词模板，提高报告质量
    3. 添加报告后处理（格式化、去除重复等）
    4. 考虑添加批量推理支持
    5. 添加推理缓存机制
    """

    def __init__(self):
        # 【TODO】这里可以初始化模型或加载提示词模板
        self.prompt_template = """
作为一名专业的放射科医生，请仔细分析这张胸部X光片，并生成详细的医学报告。

请包含以下内容：
1. 影像观察（Image Findings）
2. 病变描述（Pathology Description）
3. 初步诊断（Impression）
4. 建议（Recommendations）

用户提示：{user_prompt}
"""

    async def generate_report(
        self,
        image_path: str,
        prompt: str
    ) -> Tuple[str, float]:
        """
        生成医学报告

        参数:
            image_path: 图片路径
            prompt: 用户自定义提示词

        返回:
            (report, processing_time): 报告文本和处理时间

        【TODO】后端团队成员实现:
        ```python
        from transformers import LlavaForConditionalGeneration, AutoProcessor
        from PIL import Image

        # 1. 加载图像
        image = Image.open(image_path)

        # 2. 准备输入
        model = model_manager.get_llava_model()
        processor = AutoProcessor.from_pretrained(settings.LLAVA_MODEL_PATH)

        full_prompt = self.prompt_template.format(user_prompt=prompt)
        inputs = processor(text=full_prompt, images=image, return_tensors="pt")

        # 3. 模型推理
        start_time = time.time()
        with torch.no_grad():
            outputs = model.generate(**inputs, max_new_tokens=512)
        processing_time = time.time() - start_time

        # 4. 解码输出
        report = processor.decode(outputs[0], skip_special_tokens=True)

        return report, processing_time
        ```
        """
        logger.warning("⚠️  LLaVA报告生成逻辑待实现 (llava_service.py:71)")

        # ============ 以下是模拟代码，供前端调试使用 ============
        # 【后端团队成员需要替换为真实实现】

        # 检查图片是否存在
        if not os.path.exists(image_path):
            raise FileNotFoundError(f"图片文件不存在: {image_path}")

        # 模拟推理时间
        start_time = time.time()
        await self._simulate_inference()
        processing_time = time.time() - start_time

        # 模拟生成的报告（【TODO】替换为真实模型输出）
        mock_report = f"""
**胸部X光片分析报告**

【影像观察】
- 心影大小：轻度增大，心胸比约0.55（正常<0.5）
- 肺野：双肺纹理增多，可见肺淤血征象
- 肋膈角：双侧肋膈角清晰，未见积液
- 骨骼结构：未见明显骨折或骨质破坏

【病变描述】
1. 心脏肥大（Cardiomegaly）：心影轮廓扩大，提示可能存在心功能不全
2. 肺水肿（Pulmonary Edema）：双肺可见血管影模糊，符合肺淤血表现

【初步诊断】
1. 心脏肥大
2. 轻度肺水肿

【建议】
1. 建议进一步行心脏超声检查，评估心功能
2. 完善血常规、BNP等检查
3. 建议心内科会诊

*报告生成时间: {processing_time:.2f}秒*
*本报告由AI辅助生成，仅供参考，最终诊断请以医生判断为准*

---
用户提示词: {prompt}
        """

        logger.info(f"📝 模拟生成报告，耗时 {processing_time:.2f}秒")
        return mock_report.strip(), processing_time

    async def _simulate_inference(self):
        """模拟推理耗时"""
        import asyncio
        await asyncio.sleep(1.5)  # 模拟LLaVA推理耗时
